+++ lib/rnes/ppu.rb
@@ -238,8 +238,7 @@ module Rnes
     #      `-------- vertical flip
     def draw_sprites
       base_pattern_table_address = base_sprite_pattern_table_address
-      SPRITES_COUNT.times do |i|
-        base_sprite_ram_address = i * 4
+      0.step(SPRITES_COUNT - 1, 4) do |base_sprite_ram_address|
         y_for_sprite = (read_from_sprite_ram(base_sprite_ram_address) - TILE_HEIGHT)
         next if y_for_sprite.negative?
         name_table_index = read_from_sprite_ram(base_sprite_ram_address + 1)