+++ lib/rnes/ppu.rb
@@ -253,11 +253,9 @@ module Rnes
       0.step(SPRITES_COUNT - 1, 4) do |base_sprite_ram_address|
         y_for_sprite = (read_from_sprite_ram(base_sprite_ram_address) - TILE_HEIGHT)
         next if y_for_sprite.negative?
-        name_table_index = read_from_sprite_ram(base_sprite_ram_address + 1)
+        pattern_index = read_from_sprite_ram(base_sprite_ram_address + 1)
         sprite_attribute_byte = read_from_sprite_ram(base_sprite_ram_address + 2)
         x_for_sprite = read_from_sprite_ram(base_sprite_ram_address + 3)
-
-        pattern_index = read_pattern_index(name_table_index)
 
         mini_palette_id = sprite_attribute_byte & 0b11
         reversed_horizontally = sprite_attribute_byte[6] == 1