+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - Add CLI options and help message.
+- Add PPU sprite reverse.
 
 ## 0.1.1 - 2018-11-12