+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 - Fix PPU sprites rendering.
 - Fix PPU sprite pattern addressing.
+- Fix PPU background pattern attribute addressing.
 
 ## 0.1.1 - 2018-11-12